requirements.txt
streamlit==1.41.1
folium
streamlit-folium
pandas
fpdf2
requests
python-dotenv
plotly
